% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpit_norm.R
\name{cpit_norm}
\alias{cpit_norm}
\title{conditional PIT values for a normal distribution}
\usage{
cpit_norm(y, mean = 0, sd = 1, a = -Inf, b = Inf)
}
\arguments{
\item{y}{vector of observations.}

\item{mean}{vector of location parameters.}

\item{sd}{vector of scale parameters.}

\item{a, b}{lower/upper threshold used in the weight function.}
}
\value{
vector of conditional PIT values.
}
\description{
conditional PIT values for a normal distribution
}
\examples{
cpit_norm(y = rnorm(10))
cpit_norm(y = rnorm(10), a = 0.5)
cpit_norm(y = runif(10, 0, 5), mean = 2, sd = 0.5, a = 1.5, b = 2.5)
}
