% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpit_sample.R
\name{cpit_sample}
\alias{cpit_sample}
\title{conditional PIT values for simulated forecast distributions}
\usage{
cpit_sample(y, dat, a = -Inf, b = Inf, bw = NULL)
}
\arguments{
\item{y}{vector of observations.}

\item{dat}{matrix of simulation draws from forecast distribution.}

\item{a, b}{lower/upper threshold used in the weight function.}

\item{bw}{vector of bandwidth parameters to use in kernel density estimation.}
}
\value{
vector of conditional PIT values.
}
\description{
Using kernel density estimation
}
\examples{
cpit_sample(y = rnorm(10), dat = matrix(rnorm(50), nrow = 10))
cpit_sample(y = rnorm(10), dat = matrix(rnorm(50), nrow = 10), a = 0.5)
cpit_sample(y = runif(10, 0, 5), dat = matrix(rnorm(50, 3), nrow = 10), a = 0, b = 2.5)
}
