% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pit_hist.R
\name{pit_hist}
\alias{pit_hist}
\title{plot rank and probability integral transform (PIT) histograms}
\usage{
pit_hist(ranks, bins = NULL, pitvals = FALSE, title = NULL, ymax = NULL)
}
\arguments{
\item{ranks}{vector of ranks.}

\item{bins}{number of bins to use in the histogram.}

\item{pitvals}{logical specifying whether the input are PIT values.}

\item{title}{optional title of plot.}

\item{ymax}{optional upper limit of the y axis.}
}
\description{
plot rank and probability integral transform (PIT) histograms
}
\examples{
pit_hist(ranks = sample(1:10, 1000, replace = TRUE), title = "Example 1", ymax = 1)
pit_hist(ranks = runif(1000), bins = 10, pitvals = TRUE, title = "Example 2", ymax = 0.3)
pit_hist(ranks = rbeta(1000, shape1 = 0.5, shape2 = 0.5), pitvals = TRUE, title = "Example 3")
pit_hist(ranks = rbeta(1000, shape1 = 2, shape2 = 2), bins = 20, pitvals = TRUE, title = "Example 4")
}
