% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pit_reldiag.R
\name{pit_reldiag}
\alias{pit_reldiag}
\title{PIT reliability diagrams (adapted from https://github.com/resinj/replication_GR21)}
\usage{
pit_reldiag(
  pit,
  resampling = TRUE,
  n_resamples = 1000,
  region_level = 0.9,
  title = NULL
)
}
\arguments{
\item{pit}{vector of pit values.}

\item{resampling}{logical specifying whether resampling should be used to calculate consistency regions}

\item{n_resamples}{number of resamples to use when calculating consistency regions.}

\item{region_level}{significance level of the consistency regions.}

\item{title}{optional title of plot.}
}
\description{
PIT reliability diagrams (adapted from https://github.com/resinj/replication_GR21)
}
\examples{
pit_reldiag(pit = runif(1000), title = "Example 1")
pit_reldiag(pit = rbeta(1000, shape1 = 0.5, shape2 = 0.5), title = "Example 2")
pit_reldiag(pit = rbeta(1000, shape1 = 0.5, shape2 = 0.5), region_level = 0.5, title = "Example 3")
pit_reldiag(pit = rbeta(1000, shape1 = 2, shape2 = 2), resampling = FALSE, title = "Example 4")
}
