% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_index.R
\name{rel_index}
\alias{rel_index}
\title{function to calculate reliability index from ranks}
\usage{
rel_index(ranks, n_bins = NULL)
}
\arguments{
\item{ranks}{vector of ranks.}
}
\value{
vector of the reliability index values.
}
\description{
function to calculate reliability index from ranks
}
\examples{
rel_index(ranks = sample(1:10, 100, replace = TRUE))
}
