% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_reldiag.R
\name{tc_reldiag}
\alias{tc_reldiag}
\title{Threshold calibration reliability diagrams}
\usage{
tc_reldiag(
  x,
  y,
  t_vec,
  title = NULL,
  xlab = "x",
  ylab = "x_rc",
  pointSize = NULL,
  textSize = NULL,
  spaceLegend = NULL
)
}
\arguments{
\item{x}{matrix of probability forecasts, with columns corresponding to different thresholds.}

\item{y}{vector of observations. Length should be the same as the number of rows in \code{x}.}

\item{t_vec}{vector of thresholds.}

\item{title}{optional title.}

\item{xlab, y_lab}{optional labels of x- and y-axes.}

\item{pointSize, textSize, spaceLegend}{optional arguments for the plot legend.}
}
\value{
ggplot object containing the threshold calibration plot
}
\description{
Diagnostic plots to assess threshold calibration. These are reliability diagrams
when predicting whether multiple thresholds will be exceeded.
}
\details{
Probabilistic forecasts are threshold calibrated if they issue calibrated forecasts
for the outcome exceeding all possible thresholds. That is, a (random) predictive
distribution function $F$ is threshold calibrated if (almost surely)
\deqn{P(Y \leq y \mid F(y)) = F(y)}
for all real values \eqn{y}. See references below.

This can be assessed by choosing multiple thresholds \eqn{t} and plotting standard
reliability diagrams to check the conditional calibration of \eqn{F(t)}. These
reliability diagrams are then overlaid on one plot, called a threshold calibration
plot. If the forecasts are threshold calibrated, then all curves should be close
to the diagonal line.

\code{t_vec} contains the thresholds \eqn{t} at which calibration will be assessed.

\code{x} contains the probability forecasts \eqn{F(t)} at each of these thresholds.
\code{x} should therefore be a matrix with the number of columns equal to the length
of the vector \code{t_vec}.

\code{y} contains the corresponding observations, from which the conditional probabilities
\eqn{P(Y \leq y \mid F(y)) = F(y)} can be calculated. The length of \code{y} should be
equal to the number of rows of \code{x}.

A plot title can be added using \code{title}, and x- and y-axis labels can be
specified using \code{xlab} and \code{ylab} respectively. \code{pointSize},
\code{textSize}, and \code{spaceLegend} can be used to change the size and spacing
of the legend.
}
\examples{
n <- 1000
mu <- rnorm(n)
y <- rnorm(n, mu)

t_vec <- -2:2

# threshold calibrated
x <- sapply(t_vec, pnorm, mean = mu)
tc_reldiag(x, y, t_vec, title = "A calibrated example")

# not threshold calibrated
x <- sapply(t_vec, pnorm, mean = mu, sd = 2)
tc_reldiag(x, y, t_vec, title = "A miscalibrated example")


}
\references{
Gneiting, T., and Resin, J. (2021):
`Regression diagnostics meets forecast evaluation: Conditional calibration, reliability diagrams, and coefficient of determination'.
\emph{arXiv preprint.}
\doi{arXiv:2108.03210}
}
\author{
Sam Allen
}
