% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcrps_norm.R
\name{twcrps_norm}
\alias{twcrps_norm}
\alias{owcrps_norm}
\alias{vrcrps_norm}
\title{weighted CRPS for a normal distribution}
\usage{
twcrps_norm(y, mean = 0, sd = 1, a = -Inf, b = Inf)

owcrps_norm(y, mean = 0, sd = 1, a = -Inf, b = Inf, BS = T)

vrcrps_norm(y, mean = 0, sd = 1, a = -Inf, b = -Inf)
}
\arguments{
\item{y}{vector of observations.}

\item{mean}{vector of location parameters.}

\item{sd}{vector of scale parameters.}

\item{a, b}{lower/upper threshold used in the weight function.}

\item{BS}{logical specifying whether the outcome-weighted score should be complemented with the Brier score.}
}
\value{
vector of weighted score values.
}
\description{
weighted CRPS for a normal distribution
}
\examples{
twcrps_norm(y = rnorm(10))
owcrps_norm(y = rnorm(10))
vrcrps_norm(y = rnorm(10))
}
