% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcrps_param.R
\name{wcrps_param}
\alias{wcrps_param}
\alias{twcrps_logis}
\alias{owcrps_logis}
\alias{twcrps_norm}
\alias{owcrps_norm}
\alias{twcrps_t}
\alias{owcrps_t}
\alias{vrcrps_norm}
\title{weighted CRPS for a normal distribution}
\usage{
twcrps_logis(y, location = 0, scale = 1, a = -Inf, b = Inf)

owcrps_logis(y, location = 0, scale = 1, a = -Inf, b = Inf, BS = T)

twcrps_norm(y, mean = 0, sd = 1, a = -Inf, b = Inf)

owcrps_norm(y, mean = 0, sd = 1, a = -Inf, b = Inf, BS = T)

twcrps_t(y, df, ncp = 0, a = -Inf, b = Inf)

owcrps_t(y, df, ncp = 0, a = -Inf, b = Inf, BS = T)

vrcrps_norm(y, mean = 0, sd = 1, a = -Inf, b = -Inf)
}
\arguments{
\item{y}{vector of observations.}

\item{a, b}{numeric; lower and upper threshold defining the interval of outcomes that are of interest.}

\item{BS}{logical specifying whether the outcome-weighted scores should be complemented with the Brier score.}

\item{mean, sd, location, scale, df, ncp}{vector of parameters corresponding
to the parametric predictive distributions.}
}
\value{
vector of weighted score values.
}
\description{
weighted CRPS for a normal distribution
}
\details{
Weighted scoring rules allow forecasts to be assessed and compared when
predicting certain outcomes of interest. Weighted scoring rules are available
in the \pkg{scoringRules} package, but only for forecasts in the form of predictive samples.
Here, a collection of weighted scoring rules is provided for familiar parametric distributions.
Currently, this is available for the normal, logistic, and Student's t distributions,
though more could be added in the future.

The weight function used in the weighted scores is \code{w(z) = 1{a < z < b}}, where
\code{a} and \code{b} are numerical values corresponding to the lower and upper
bound of the interval of interest. The default values are \code{a = -Inf} and
\code{b = Inf}, which returns the unweighted CRPS. Future work could also integrate
the option to employ custom weight functions, as in \pkg{scoringRules}.

The parameter arguments (\code{mean}, \code{sd}, \code{location}, \code{scale}, \code{df}, \code{ncp})
should be vectors of the same length as the observation vector \code{y}. The naming of these parameters
is the same as in the standard \proglang{R} distribution handles, e.g. \code{pnorm}.

Threshold-weighted and outcome-weighted scoring rules are provided. The vertically re-scaled CRPS
is also provided for forecasts in the form of a normal distribution. This could be
added for other distributions in the future.
}
\examples{
sig <- 1
mu <- rnorm(20, mean = 0, sd = 5)
y <- rnorm(20, mean = mu, sd = sig)

scoringRules::crps_norm(y = y, mean = mu, sd = sig)
twcrps_norm(y = y, mean = mu, sd = sig)
owcrps_norm(y = y, mean = mu, sd = sig)
vrcrps_norm(y = y, mean = mu, sd = sig)

# emphasise outcomes above 0
twcrps_logis(y = y, location = mu, scale = sig, a = 0)
owcrps_logis(y = y, location = mu, scale = sig, a = 0)

# emphasise outcomes between -1 and 1
twcrps_t(y = y, df = 5, ncp = mu, a = -1, b = 1)
owcrps_t(y = y, df = 5, ncp = mu, a = -1, b = 1)

}
\references{
\emph{Threshold-weighted CRPS:}

Gneiting, T. and R. Ranjan (2011):
`Comparing density forecasts using threshold-and quantile-weighted scoring rules',
\emph{Journal of Business & Economic Statistics} 29, 411-422.
\doi{10.1198/jbes.2010.08110}

\emph{Outcome-weighted CRPS:}

Holzmann, H. and B. Klar (2017):
`Focusing on regions of interest in forecast evaluation',
\emph{Annals of Applied Statistics} 11, 2404-2431.
\doi{10.1214/17-AOAS1088}

\emph{Vertically re-scaled CRPS:}

Allen, S., Ginsbourger, D. and J. Ziegel (2022):
`Evaluating forecasts for high-impact events using transformed kernel scores',
\emph{arXiv preprint} arXiv:2202.12732.
\doi{10.48550/arXiv.2202.12732}
}
\author{
Sam Allen
}
